scraper/requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
asyncpg==0.30.0
python-dotenv==1.0.0
redis==4.5.4
SQLAlchemy==2.0.36
pydantic==2.9.0
pydantic-settings==2.6.0
greenlet==3.0.1
